# feat(lanelet2_extension): copy the z values in getExpandedLanelet