--- a/tmp/lanelet2_extension/include/lanelet2_extension/utility/utilities.hpp
+++ b/tmp/lanelet2_extension/include/lanelet2_extension/utility/utilities.hpp
@@ -1,4 +1,4 @@
-// Copyright 2015-2019 Autoware Foundation. All rights reserved.
+// Copyright 2015-2023 Autoware Foundation. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.