--- a/tmp/lanelet2_extension/lib/utilities.cpp
+++ b/tmp/lanelet2_extension/lib/utilities.cpp
@@ -444,28 +444,9 @@ lanelet::ConstLanelet getExpandedLanelet(
 
   lanelet::Points3d ex_lefts = toPoints3d(expanded_left_bound_2d);
   lanelet::Points3d ex_rights = toPoints3d(expanded_right_bound_2d);
+  copyZ(lanelet_obj.leftBound3d(), ex_lefts);
+  copyZ(lanelet_obj.rightBound3d(), ex_rights);
 
-  const auto copy_z = [](const lanelet::ConstLineString3d & from, lanelet::Points3d & to) {
-    if (from.empty() || to.empty()) return;
-    to.front().z() = from.front().z();
-    if (from.size() < 2 || to.size() < 2) return;
-    to.back().z() = from.back().z();
-    auto i_from = 0lu;
-    auto s_from = 0.0;
-    auto s_to = 0.0;
-    auto s_from_prev = 0.0;
-    for (auto i_to = 1lu; i_to + 1 < to.size(); ++i_to) {
-      s_to += lanelet::geometry::distance2d(to[i_to - 1], to[i_to]);
-      for (; s_from < s_to && i_from + 1 < from.size(); ++i_from) {
-        s_from_prev = s_from;
-        s_from += lanelet::geometry::distance2d(from[i_from], from[i_from + 1]);
-      }
-      const auto ratio = (s_to - s_from_prev) / (s_from - s_from_prev);
-      to[i_to].z() = from[i_from - 1].z() + ratio * (from[i_from].z() - from[i_from - 1].z());
-    }
-  };
-  copy_z(lanelet_obj.leftBound3d(), ex_lefts);
-  copy_z(lanelet_obj.rightBound3d(), ex_rights);
   const auto & extended_left_bound_3d = lanelet::LineString3d(lanelet::InvalId, ex_lefts);
   const auto & expanded_right_bound_3d = lanelet::LineString3d(lanelet::InvalId, ex_rights);
   const auto & lanelet = lanelet::Lanelet(