--- a/tmp/lanelet2_extension/test/src/test_utilities.cpp
+++ b/tmp/lanelet2_extension/test/src/test_utilities.cpp
@@ -114,7 +114,7 @@ TEST_F(TestSuite, OverwriteLaneletsCenterline)  // NOLINT for gtest
   }
 }
 
-TEST(Utilities, copyZ)
+TEST(Utilities, copyZ)  // NOLINT for gtest
 {
   using lanelet::utils::copyZ;