Signed-off-by: Maxime CLEMENT <maxime.clement@tier4.jp>

## Description

<!-- Write a brief description of this PR. -->
Update function `getExpandedLanelet` to set correct `z` values in the calculated expended lanelet.

| Without this PR | With this PR |
|:-:|:-:|
| ![image](https://user-images.githubusercontent.com/78338830/211228870-2995bb49-40ea-4c69-a114-44b5bd3adfd9.png) | ![image](https://user-images.githubusercontent.com/78338830/211228767-2fb66dc6-6a97-49a1-b380-dfd8568653d4.png) |

After some quick profiling, it seems the added code takes at most 4μs so this should not impact performances.

### How to test

You need to configure the `behavior_path_planner` to extend the lanelets by modifying `autoware_launch/config/planning/scenario_planning/lane_driving/behavior_planning/behavior_path_planner/drivable_area_expansion.param.yaml` and use non-zero parameter values (e.g., for the `lane_following`).

## Pre-review checklist for the PR author

The PR author **must** check the checkboxes below when creating the PR.

- [X] I've confirmed the [contribution guidelines].
- [X] The PR follows the [pull request guidelines].

## In-review checklist for the PR reviewers

The PR reviewers **must** check the checkboxes below before approval.

- [x] The PR follows the [pull request guidelines].

## Post-review checklist for the PR author

The PR author **must** check the checkboxes below before merging.

- [x] There are no open discussions or they are tracked via tickets.

After all checkboxes are checked, anyone who has write access can merge the PR.

[contribution guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/
[pull request guidelines]: https://autowarefoundation.github.io/autoware-documentation/main/contributing/pull-request-guidelines/