@mitsudome-r 2023-01-16
The change looks go to me. I want some feedback from @kosuke55 @purewater0901 (or anyone else from TIER IV planning team) about if it is okay to copy the z value in expanded lanelets in first place. (i.e., are there any codes that assumes flat lanelet after expansion?)